Moved metrics into separate package
