+++ metrics/stats.go
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package api
+package metrics
 
 import (
 	"github.com/onitake/restreamer/protocol"
